+++ src/op/mod.rs
@@ -124,8 +124,8 @@ pub mod is_simple;
 pub mod is_symmetric;
 pub mod iter_all_arcs;
 pub mod iter_all_weighted_arcs;
-pub mod iter_arcs;
-pub mod iter_arcs_mut;
+pub mod iter_out_neighbors;
+pub mod iter_out_neighbors_mut;
 pub mod iter_vertices;
 pub mod iter_weighted_arcs;
 pub mod order;
@@ -149,8 +149,8 @@ pub use {
     is_symmetric::IsSymmetric,
     iter_all_arcs::IterAllArcs,
     iter_all_weighted_arcs::IterAllWeightedArcs,
-    iter_arcs::IterArcs,
-    iter_arcs_mut::IterArcsMut,
+    iter_out_neighbors::IterOutNeighbors,
+    iter_out_neighbors_mut::IterOutNeighborsMut,
     iter_vertices::IterVertices,
     iter_weighted_arcs::IterWeightedArcs,
     order::Order,